+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Build docker images for ARM64 architecture
 - New option to `assembly upgrade` command:
    - `--start` - automatically start server if upgrade is successful and server was not running before upgrade.
+- Added new option `odood assembly --assembly-path` that could be used to specify different assembly path for assembly commands.
+  Mostly, this option could be useful for CI
 
 ### Changed